Move dagster_dg.cli.shared_options to dagster_dg.shared_options
Summary:
This is the only module in dagster_dg.cli that create_dagster uses. Moving it out of that folder lets us do a clean lift of dagster_dg.cli into a separate dagster_dg_cli package.

BK

> Insert changelog entry or delete this section.
