+++ examples/docs_snippets/docs_snippets_tests/snippet_checks/guides/components/integrations/test_dlt_component.py
@@ -2,7 +2,7 @@ import textwrap
 from contextlib import ExitStack
 from pathlib import Path
 
-from dagster_dg.cli.utils import activate_venv, environ
+from dagster_dg.utils import activate_venv
 
 from dagster._utils.env import environ
 from docs_snippets_tests.snippet_checks.guides.components.utils import (
