+++ examples/docs_snippets/docs_snippets_tests/snippet_checks/guides/components/test_component_docs_adding_attributes_to_assets.py
@@ -2,9 +2,8 @@ import textwrap
 from contextlib import ExitStack
 from pathlib import Path
 
-from dagster_dg.cli.utils import activate_venv, environ
+from dagster_dg.utils import activate_venv
 
-from dagster._utils.env import environ
 from docs_snippets_tests.snippet_checks.guides.components.utils import DAGSTER_ROOT
 from docs_snippets_tests.snippet_checks.utils import (
     _run_command,
