+++ examples/docs_snippets/docs_snippets_tests/snippet_checks/guides/components/test_components_docs.py
@@ -2,11 +2,10 @@ import os
 import textwrap
 from contextlib import ExitStack
 from pathlib import Path
-from typing import Literal, Optional
 
 import pytest
-from dagster_dg.cli.utils import activate_venv, environ
-from typing_extensions import TypeAlias
+from dagster_dg.utils import activate_venv
+from dagster_shared.utils import environ
 
 from docs_snippets_tests.snippet_checks.guides.components.utils import (
     DAGSTER_ROOT,
