+++ examples/docs_snippets/docs_snippets_tests/snippet_checks/guides/components/test_components_docs_creating_a_component.py
@@ -1,7 +1,7 @@
 from contextlib import ExitStack
 from pathlib import Path
 
-from dagster_dg.cli.utils import activate_venv
+from dagster_dg.utils import activate_venv
 
 from docs_snippets_tests.snippet_checks.guides.components.utils import (
     DAGSTER_ROOT,
