+++ examples/docs_snippets/docs_snippets_tests/snippet_checks/guides/dg/test_migrating_project/test_migrating_project.py
@@ -3,7 +3,7 @@ from contextlib import ExitStack
 from pathlib import Path
 
 import pytest
-from dagster_dg.cli.utils import activate_venv
+from dagster_dg.utils import activate_venv
 
 from docs_snippets_tests.snippet_checks.guides.components.test_components_docs import (
     DgTestPackageManager,
