+++ python_modules/libraries/create-dagster/create_dagster_tests/test_scaffold_commands.py
@@ -6,7 +6,7 @@ from typing import Literal, Optional, get_args
 import dagster_shared.check as check
 import pytest
 import tomlkit
-from dagster_dg.cli.shared_options import DEFAULT_EDITABLE_DAGSTER_PROJECTS_ENV_VAR
+from dagster_dg.shared_options import DEFAULT_EDITABLE_DAGSTER_PROJECTS_ENV_VAR
 from dagster_dg.utils import (
     create_toml_node,
     discover_git_root,
