+++ python_modules/libraries/dagster-dg/dagster_dg/cli/dev.py
@@ -6,10 +6,10 @@ from typing import Optional, TypeVar
 
 import click
 
-from dagster_dg.cli.shared_options import dg_global_options, dg_path_options
 from dagster_dg.cli.utils import create_temp_workspace_file
 from dagster_dg.config import normalize_cli_config
 from dagster_dg.context import DgContext
+from dagster_dg.shared_options import dg_global_options, dg_path_options
 from dagster_dg.utils import DgClickCommand, pushd, validate_dagster_availability
 from dagster_dg.utils.telemetry import cli_telemetry_wrapper
