+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_launch_commands.py
@@ -9,7 +9,7 @@ import yaml
 from dagster_dg.utils import ensure_dagster_dg_tests_import
 
 ensure_dagster_dg_tests_import()
-from dagster_dg.cli.utils import activate_venv
+from dagster_dg.utils import activate_venv
 
 from dagster_dg_tests.utils import ProxyRunner, isolated_example_project_foo_bar
